fix(aws-api-mcp-server): Add profile support to boto3 operations (#986)
* Add profile support to boto3 operations

* Update CHANGELOG
