+++ src/aws-api-mcp-server/tests/aws/test_service.py
@@ -4,7 +4,6 @@ from ..history_handler import history
 from awslabs.aws_api_mcp_server.core.aws.driver import translate_cli_to_ir
 from awslabs.aws_api_mcp_server.core.aws.service import (
     execute_awscli_customization,
-    get_local_credentials,
     interpret_command,
     is_operation_read_only,
     validate,
@@ -15,7 +14,6 @@ from awslabs.aws_api_mcp_server.core.common.models import (
     AwsCliAliasResponse,
     CommandMetadata,
     Context,
-    Credentials,
     InterpretationMetadata,
     InterpretationResponse,
     IRTranslation,
@@ -24,7 +22,6 @@ from awslabs.aws_api_mcp_server.core.common.models import (
 )
 from awslabs.aws_api_mcp_server.core.metadata.read_only_operations_list import ReadOnlyOperations
 from botocore.config import Config
-from botocore.exceptions import NoCredentialsError
 from tests.fixtures import (
     CLOUD9_DESCRIBE_ENVIRONMENTS,
     CLOUD9_LIST_ENVIRONMENTS,
@@ -36,7 +33,6 @@ from tests.fixtures import (
     GET_CALLER_IDENTITY_PAYLOAD,
     SSM_LIST_NODES_PAYLOAD,
     T2_EC2_DESCRIBE_INSTANCES_FILTERED,
-    TEST_CREDENTIALS,
     patch_boto3,
 )
 from typing import Any
@@ -56,10 +52,8 @@ from unittest.mock import MagicMock, patch
 )
 def test_interpret_returns_validation_failures(cli_command, reason, service, operation):
     """Test that interpret_command returns validation failures for invalid operations."""
-    credentials = Credentials(**TEST_CREDENTIALS)
     response = interpret_command(
         cli_command=cli_command,
-        credentials=credentials,
         default_region='us-east-1',
     )
     assert response.response is None
@@ -80,10 +74,8 @@ def test_interpret_returns_validation_failures(cli_command, reason, service, ope
 
 def test_interpret_returns_missing_context_failures():
     """Test that interpret_command returns missing context failures when required parameters are missing."""
-    credentials = Credentials(**TEST_CREDENTIALS)
     response = interpret_command(
         cli_command=CLOUD9_PARAMS_CLI_MISSING_CONTEXT,
-        credentials=credentials,
         default_region='us-east-1',
     )
     assert response.response is None
@@ -201,10 +193,7 @@ def test_interpret_returns_valid_response(
     """Test that interpret_command returns a valid response for correct CLI commands."""
     with patch_boto3():
         history.events.clear()
-        credentials = Credentials(**TEST_CREDENTIALS)
-        response = interpret_command(
-            cli_command=cli, default_region='us-east-1', credentials=credentials
-        )
+        response = interpret_command(cli_command=cli, default_region='us-east-1')
         assert response == ProgramInterpretationResponse(
             response=InterpretationResponse(json=as_json(output), error=None, status_code=200),
             failed_constraints=[],
@@ -226,10 +215,8 @@ def test_interpret_injects_region():
         with patch('awslabs.aws_api_mcp_server.core.parser.interpretation.Config') as patch_config:
             history.events.clear()
             patch_config.return_value = default_config
-            credentials = Credentials(**TEST_CREDENTIALS)
             response = interpret_command(
                 cli_command='aws cloud9 describe-environments --environment-ids 7d61007bd98b4d589f1504af84c168de b181ffd35fe2457c8c5ae9d75edc068a',
-                credentials=credentials,
                 default_region=region,
             )
             assert response.metadata == InterpretationMetadata(
@@ -273,11 +260,9 @@ def test_interpret_injects_region():
 def test_region_picked_up_from_arn(cli, region):
     """Test that region is correctly picked up from ARN in the CLI command."""
     with patch_boto3():
-        credentials = Credentials(**TEST_CREDENTIALS)
         response = interpret_command(
             cli_command=cli,
             default_region='us-east-1',
-            credentials=credentials,
         )
         assert response.metadata is not None
         assert response.metadata.region_name == region
@@ -453,67 +438,6 @@ def test_is_operation_read_only_raises_error_for_missing_operation_name():
 
     with pytest.raises(RuntimeError, match='failed to check if operation is allowed'):
         is_operation_read_only(ir, read_only_operations)
-
-
-@patch('awslabs.aws_api_mcp_server.core.aws.service.AWS_API_MCP_PROFILE_NAME', 'test')
-@patch('awslabs.aws_api_mcp_server.core.aws.service.boto3.Session')
-def test_get_local_credentials_success_with_aws_mcp_profile(mock_session_class):
-    """Test get_local_credentials returns credentials when available."""
-    mock_session = MagicMock()
-    mock_session_class.return_value = mock_session
-
-    mock_credentials = MagicMock()
-    mock_credentials.access_key = 'test-access-key'
-    mock_credentials.secret_key = 'test-secret-key'  # pragma: allowlist secret
-    mock_credentials.token = 'test-session-token'
-
-    mock_session.get_credentials.return_value = mock_credentials
-
-    result = get_local_credentials()
-
-    assert isinstance(result, Credentials)
-    assert result.access_key_id == 'test-access-key'
-    assert result.secret_access_key == 'test-secret-key'  # pragma: allowlist secret
-    assert result.session_token == 'test-session-token'
-    mock_session_class.assert_called_once_with(profile_name='test')
-    mock_session.get_credentials.assert_called_once()
-
-
-@patch('awslabs.aws_api_mcp_server.core.aws.service.boto3.Session')
-def test_get_local_credentials_success_with_default_creds(mock_session_class):
-    """Test get_local_credentials returns credentials when available."""
-    mock_session = MagicMock()
-    mock_session_class.return_value = mock_session
-
-    mock_credentials = MagicMock()
-    mock_credentials.access_key = 'test-access-key'
-    mock_credentials.secret_key = 'test-secret-key'  # pragma: allowlist secret
-    mock_credentials.token = 'test-session-token'
-
-    mock_session.get_credentials.return_value = mock_credentials
-
-    result = get_local_credentials()
-
-    assert isinstance(result, Credentials)
-    assert result.access_key_id == 'test-access-key'
-    assert result.secret_access_key == 'test-secret-key'  # pragma: allowlist secret
-    assert result.session_token == 'test-session-token'
-    mock_session_class.assert_called_once()
-    mock_session.get_credentials.assert_called_once()
-
-
-@patch('awslabs.aws_api_mcp_server.core.aws.service.boto3.Session')
-def test_get_local_credentials_raises_no_credentials_error(mock_session_class):
-    """Test get_local_credentials raises NoCredentialsError when credentials are None."""
-    mock_session = MagicMock()
-    mock_session_class.return_value = mock_session
-    mock_session.get_credentials.return_value = None
-
-    with pytest.raises(NoCredentialsError):
-        get_local_credentials()
-
-    mock_session_class.assert_called_once()
-    mock_session.get_credentials.assert_called_once()
 
 
 @patch('awslabs.aws_api_mcp_server.core.aws.service.driver')
